+++ examples/complete/main.tf
@@ -16,23 +16,9 @@ provider "aws" {
 }
 
 locals {
-  name   = "example-asg"
+  name   = "ex-asg-complete"
   region = "eu-west-1"
-
-  tags = [
-    {
-      key                 = "Project"
-      value               = "megasecret"
-      propagate_at_launch = true
-    },
-    {
-      key                 = "foo"
-      value               = "something"
-      propagate_at_launch = true
-    },
-  ]
-
-  tags_as_map = {
+  tags = {
     Owner       = "user"
     Environment = "dev"
   }
@@ -61,7 +47,7 @@ module "vpc" {
   enable_dns_hostnames = true
   enable_dns_support   = true
 
-  tags = local.tags_as_map
+  tags = local.tags
 }
 
 module "asg_sg" {
@@ -82,7 +68,7 @@ module "asg_sg" {
 
   egress_rules = ["all-all"]
 
-  tags = local.tags_as_map
+  tags = local.tags
 }
 
 data "aws_ami" "amazon_linux" {
@@ -112,28 +98,26 @@ resource "aws_iam_service_linked_role" "autoscaling" {
 resource "aws_iam_instance_profile" "ssm" {
   name = "complete-${local.name}"
   role = aws_iam_role.ssm.name
-  tags = local.tags_as_map
+  tags = local.tags
 }
 
 resource "aws_iam_role" "ssm" {
   name = "complete-${local.name}"
-  tags = local.tags_as_map
+  tags = local.tags
 
-  assume_role_policy = <<-EOT
-  {
-    "Version": "2012-10-17",
-    "Statement": [
+  assume_role_policy = jsonencode({
+    Version = "2012-10-17",
+    Statement = [
       {
-        "Action": "sts:AssumeRole",
-        "Principal": {
-          "Service": "ec2.amazonaws.com"
+        Action = "sts:AssumeRole",
+        Principal = {
+          Service = "ec2.amazonaws.com"
         },
-        "Effect": "Allow",
-        "Sid": ""
+        Effect = "Allow",
+        Sid    = ""
       }
     ]
-  }
-  EOT
+  })
 }
 
 module "alb_http_sg" {
@@ -146,7 +130,7 @@ module "alb_http_sg" {
 
   ingress_cidr_blocks = ["0.0.0.0/0"]
 
-  tags = local.tags_as_map
+  tags = local.tags
 }
 
 module "alb" {
@@ -176,7 +160,7 @@ module "alb" {
     },
   ]
 
-  tags = local.tags_as_map
+  tags = local.tags
 }
 
 ################################################################################
@@ -211,8 +195,7 @@ module "launch_template_only" {
   image_id      = data.aws_ami.amazon_linux.id
   instance_type = "t3.micro"
 
-  tags        = local.tags
-  tags_as_map = local.tags_as_map
+  tags = local.tags
 }
 
 ################################################################################
@@ -233,8 +216,7 @@ module "default" {
   image_id      = data.aws_ami.amazon_linux.id
   instance_type = "t3.micro"
 
-  tags        = local.tags
-  tags_as_map = local.tags_as_map
+  tags = local.tags
 }
 
 ################################################################################
@@ -266,8 +248,7 @@ module "external" {
   create_launch_template = false
   launch_template        = aws_launch_template.this.name
 
-  tags        = local.tags
-  tags_as_map = local.tags_as_map
+  tags = local.tags
 }
 
 ################################################################################
@@ -385,15 +366,13 @@ module "complete" {
 
   instance_market_options = {
     market_type = "spot"
-    spot_options = {
-      block_duration_minutes = 60
-    }
   }
 
   metadata_options = {
     http_endpoint               = "enabled"
     http_tokens                 = "required"
     http_put_response_hop_limit = 32
+    instance_metadata_tags      = "enabled"
   }
 
   network_interfaces = [
@@ -422,16 +401,15 @@ module "complete" {
     },
     {
       resource_type = "volume"
-      tags          = merge({ WhatAmI = "Volume" }, local.tags_as_map)
+      tags          = merge({ WhatAmI = "Volume" })
     },
     {
       resource_type = "spot-instances-request"
-      tags          = merge({ WhatAmI = "SpotInstanceRequest" }, local.tags_as_map)
+      tags          = merge({ WhatAmI = "SpotInstanceRequest" })
     }
   ]
 
-  tags        = local.tags
-  tags_as_map = local.tags_as_map
+  tags = local.tags
 
   # Autoscaling Schedule
   schedules = {
@@ -456,6 +434,39 @@ module "complete" {
       desired_capacity = 0
       start_time       = "2031-12-31T10:00:00Z" # Should be in the future
       end_time         = "2032-01-01T16:00:00Z"
+    }
+  }
+  # Target scaling policy schedule based on average CPU load
+  scaling_policies = {
+    avg-cpu-policy-greater-than-50 = {
+      policy_type               = "TargetTrackingScaling"
+      estimated_instance_warmup = 1200
+      target_tracking_configuration = {
+        predefined_metric_specification = {
+          predefined_metric_type = "ASGAverageCPUUtilization"
+        }
+        target_value = 50.0
+      }
+    },
+    predictive-scaling = {
+      policy_type = "PredictiveScaling"
+      predictive_scaling_configuration = {
+        mode                         = "ForecastAndScale"
+        scheduling_buffer_time       = 10
+        max_capacity_breach_behavior = "IncreaseMaxCapacity"
+        max_capacity_buffer          = 10
+        metric_specification = {
+          target_value = 32
+          predefined_scaling_metric_specification = {
+            predefined_metric_type = "ASGAverageCPUUtilization"
+            resource_label         = "testLabel"
+          }
+          predefined_load_metric_specification = {
+            predefined_metric_type = "ASGTotalCPUUtilization"
+            resource_label         = "testLabel"
+          }
+        }
+      }
     }
   }
 }
@@ -530,8 +541,7 @@ module "mixed_instance" {
     ]
   }
 
-  tags        = local.tags
-  tags_as_map = local.tags_as_map
+  tags = local.tags
 }
 
 ################################################################################
@@ -558,6 +568,5 @@ module "warm_pool" {
     max_group_prepared_capacity = 2
   }
 
-  tags        = local.tags
-  tags_as_map = local.tags_as_map
+  tags = local.tags
 }