+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file.
 
+## [5.0.0](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/compare/v4.11.0...v5.0.0) (2022-02-14)
+
+
+### ⚠ BREAKING CHANGES
+
+* Update autoscaling group `tags` -> `tag` to support v4 of AWS provider (#179)
+
+### Features
+
+* Update autoscaling group `tags` -> `tag` to support v4 of AWS provider ([#179](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/issues/179)) ([2c2a8a9](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/commit/2c2a8a9bc255b49dc63f587d090b6b5a7d8b7c94))
+
+## [4.11.0](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/compare/v4.10.0...v4.11.0) (2022-01-17)
+
+
+### Features
+
+* Add support for aws_autoscaling_policy ([#175](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/issues/175)) ([a8ba844](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/commit/a8ba844aee63309664c6cbaac582d26ce77fd655))
+
+## [4.10.0](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/compare/v4.9.0...v4.10.0) (2022-01-14)
+
+
+### Features
+
+* Add `instance_metadata_tags` attribute to launch temlplate and bump AWS provider to support ([#176](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/issues/176)) ([4019026](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/commit/40190265ef651ef89d5e7a98dcf303a4757f40cd))
+
 # [4.9.0](https://github.com/terraform-aws-modules/terraform-aws-autoscaling/compare/v4.8.1...v4.9.0) (2021-11-22)